feat: Move DoorViewModel, FCM use cases, DoorFcmRepository to shared modules (Phase 27b)
Move DoorViewModel + interface to usecase/src/commonMain/.
Move DoorFcmRepository interface to domain/src/commonMain/.
Move FCM use cases (FetchFcmStatus, RegisterFcm, DeregisterFcm) to usecase/.
Remove @Inject — DI wired via explicit @Provides in AppComponent.

Both RemoteButtonViewModel and DoorViewModel are now in shared modules.
